+++ app/Http/Controllers/LogsController.php
@@ -1,21 +1,53 @@
 <?php
 namespace App\Http\Controllers;
+use Exception;
 use App\Models\Log;
 use App\Helpers\ApiResponse;
+use App\Services\LogService;
+use Illuminate\Http\Request;
+use Illuminate\Http\JsonResponse;
+
 use App\Http\Resources\LogResource;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\CreateLogRequest;
 use App\Http\Requests\UpdateLogRequest;
-use Illuminate\Http\JsonResponse;
 
 
 class LogsController extends Controller
 {
-    public function index(): JsonResponse
+    protected $logService;
+    public function __construct(LogService $logService)
     {
-        $logs = Log::with(['room', 'user'])->paginate(10);
+        $this->logService = $logService;
+    }
 
-        return ApiResponse::paginated('Logs fetched successfully', $logs, LogResource::class);
+    public function index(Request $request): JsonResponse
+    {
+        Gate::authorize('view', Log::class);
+         $params = $request->only([
+            'search',
+            'sort_by',
+            'sort_dir',
+            'per_page',
+            'with',
+            'columns',
+            'exact'
+        ]);
+        $columns = ['name'];
+        try {
+            $items = $this->logService->searchPaginatedList($params, $columns);
+            return ApiResponse::success(
+                'Logs fetched successfully.',
+                $items,
+                200,
+                LogResource::class
+            );
+           
+        } catch (Exception $e) {
+            return ApiResponse::error('Failed to fetch logs.', 500);
+        }
+        
     }
 
     public function store(CreateLogRequest $request): JsonResponse
@@ -27,6 +59,7 @@ class LogsController extends Controller
 
     public function show(Log $log): JsonResponse
     {
+        Gate::authorize('view', Log::class);
         return ApiResponse::success('Log fetched successfully', new LogResource($log->load(['room', 'user'])));
     }
 
@@ -39,6 +72,7 @@ class LogsController extends Controller
 
     public function destroy(Log $log): JsonResponse
     {
+        Gate::authorize('delete', Log::class);
         $log->delete();
 
         return ApiResponse::success('Log deleted successfully', null, 200);
